Add visualization graph (#39)
Generate an svg graph from the core IR using pygraphviz and adding some interactivity with javascript
---------
