--- Parser_PoC/svg-interactive-style.css
@@ -1,70 +1,0 @@
-/* The first 2 styles are hiding longer description texts on */
-/* nodes and edges, that are shown when nodes are clicked */
-/* .edge text{ */
-/*     opacity: 0; */
-/* } */
-/* .node text:not(:first-of-type){ */
-/*     opacity: 0; */
-/* } */
-
-.text-highlight-nodes text{
-    opacity: 1 !important;
-    stroke-width: 10;
-    font-size: 20px;
-    font-weight: bold;
-    fill: black;
-}
-.text-highlight-edges text{
-    opacity: 1 !important;
-    stroke-width: 10;
-    font-size: 20px;
-    font-weight: bold;
-    fill: Indigo;
-}
-.edge-highlight path{
-    opacity: 1;
-    stroke-width: 10;
-    stroke: crimson;
-}
-.edge-highlight polygon{
-    opacity: 1;
-    stroke-width: 10;
-    stroke: crimson;
-}
-.node-highlight polygon{
-    opacity: 1;
-    stroke-width: 10;
-    stroke: crimson;
-    z-index:99999;
-}
-.node-highlight ellipse{
-    opacity: 1;
-    stroke-width: 10;
-    stroke: crimson;
-    z-index:99999;
-}
-.node-highlight path{
-    opacity: 1;
-    stroke-width: 10;
-    stroke: crimson;
-    z-index:99999;
-}
-.compass {
-    fill: #fff;
-    stroke: #000;
-    stroke-width: 1;
-}
-.button {
-    fill: #225EA8;
-    stroke: #0C2C84;
-    stroke-width: 0.5;
-    stroke-miterlimit:6;
-    stroke-linecap: round;
-}
-.button:hover {
-    stroke-width: 1;
-}
-.plus-minus {
-    fill: #fff;
-    pointer-events: none;
-}
