+++ src/sirocco/core.py
@@ -250,6 +250,7 @@ class Workflow:
     """Internal reprensentation of a workflow"""
 
     def __init__(self, workflow_config: ConfigWorkflow) -> None:
+        self.name = workflow_config.name
         self.tasks = Store()
         self.data = Store()
         self.cycles = Store()
