+++ src/sirocco/svg-interactive-style.css
@@ -1,0 +1,85 @@
+/* https://github.com/BartBrood/Interactive-Graphviz-Diagrams */
+
+/* MIT License */
+
+/* Copyright (c) 2024 BartBrood */
+
+/* Permission is hereby granted, free of charge, to any person obtaining a copy */
+/* of this software and associated documentation files (the "Software"), to deal */
+/* in the Software without restriction, including without limitation the rights */
+/* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell */
+/* copies of the Software, and to permit persons to whom the Software is */
+/* furnished to do so, subject to the following conditions: */
+
+/* The above copyright notice and this permission notice shall be included in all */
+/* copies or substantial portions of the Software. */
+
+/* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR */
+/* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, */
+/* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE */
+/* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER */
+/* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, */
+/* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE */
+/* SOFTWARE. */
+
+
+/* The first 2 styles are hiding longer description texts on */
+/* nodes and edges, that are shown when nodes are clicked */
+/* .edge text{ */
+/*     opacity: 0; */
+/* } */
+/* .node text:not(:first-of-type){ */
+/*     opacity: 0; */
+/* } */
+
+.text-highlight-nodes text{
+    opacity: 1 !important;
+    stroke-width: 5;
+    font-size: 20px;
+    font-weight: bold;
+    fill: black;
+}
+.text-highlight-edges text{
+    opacity: 1 !important;
+    stroke-width: 5;
+    font-size: 20px;
+    font-weight: bold;
+    fill: Indigo;
+}
+.edge-highlight path{
+    opacity: 1;
+    stroke-width: 5;
+    stroke: crimson;
+}
+.edge-highlight polygon{
+    opacity: 1;
+    stroke-width: 5;
+    stroke: crimson;
+}
+.node-highlight polygon{
+    opacity: 1;
+    stroke-width: 5;
+    stroke: crimson;
+    z-index:99999;
+}
+.node-highlight ellipse{
+    opacity: 1;
+    stroke-width: 5;
+    stroke: crimson;
+    z-index:99999;
+}
+.node-highlight path{
+    opacity: 1;
+    stroke-width: 5;
+    stroke: crimson;
+    z-index:99999;
+}
+.compass {
+    fill: #fff;
+    stroke: #000;
+    stroke-width: 1;
+}
+.plus-minus {
+    fill: #fff;
+    pointer-events: none;
+}
